FIX: DeleteMapping response status edited from 204 to 200. UPD: Update film - remove director if he doesn`t exists. UPD: Remove like - validation film_id and user_id. UPD: getFeed - validation user. FIX: ADD_LIKE SQL - remake from INSERT to MERGE

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -76,6 +76,13 @@ public class FilmService {
     }
 
     public void removeLike(Long filmId, Long userId) {
+        if (!filmRepository.findById(filmId).isPresent()) {
+            throw new NotFoundException("Фильм с id = " + filmId + " не найден");
+        }
+        if (!userRepository.findById(userId).isPresent()) {
+            throw new NotFoundException("Пользователь с id = " + userId + " не найден");
+        }
+
         filmRepository.removeLike(filmId, userId);
         eventService.addEvent(userId, EventType.LIKE, Operation.REMOVE, filmId);
     }
